Add tensor and vector math to math
